refactor loginform and signupform into one session folder
