+++ frontend/src/components/Navigation/index.js
@@ -1,13 +1,20 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { NavLink } from 'react-router-dom';
-import { useSelector } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 import ProfileButton from './ProfileButton';
-import LoginFormModal from '../LoginFormModal';
-import SignupFormModal from '../SignupFormModal';
+import {LoginFormModal, SignupFormModal} from '../SessionForms';
+import * as sessionActions from "../../store/session"
 import './Navigation.css';
 
 function Navigation(){
+  // const dispatch = useDispatch()
   const sessionUser = useSelector(state => state.session.user);
+
+  // useEffect(() => {
+  //   if (sessionUser) {
+  //     dispatch(sessionActions.restoreSession())
+  //   }
+  // },[dispatch, sessionUser])
 
   let sessionLinks;
   if (sessionUser) {
@@ -24,12 +31,14 @@ function Navigation(){
   }
 
   return (
-    <ul>
-      <li>
-        <NavLink exact to="/">Home</NavLink>
+    <header className="site-header">
+      <NavLink exact to="/" className="nav-title">
+        <h1>Venture Camp</h1>
+      </NavLink>
+      <div className="session-links">
         {sessionLinks}
-      </li>
-    </ul>
+      </div>
+    </header>
   );
 }
