+++ frontend/src/components/SessionForms/LoginForm.js
@@ -1,0 +1,82 @@
+import React, { useState } from "react";
+import * as sessionActions from "../../store/session";
+import { useDispatch } from "react-redux";
+
+function LoginForm() {
+  const dispatch = useDispatch();
+  const [credential, setCredential] = useState("");
+  const [password, setPassword] = useState("");
+  const [errors, setErrors] = useState([]);
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    setErrors([]);
+    return dispatch(sessionActions.login({ credential, password }))
+      .catch(async (res) => {
+        let data;
+        try {
+          // .clone() essentially allows you to read the response body twice
+          data = await res.clone().json();
+        } catch {
+          data = await res.text(); // Will hit this case if the server is down
+        }
+        if (data?.errors) setErrors(data.errors);
+        else if (data) setErrors([data]);
+        else setErrors([res.statusText]);
+      });
+  };
+
+  const handleDemoLogin = (e) => {
+    e.preventDefault()
+    const demoCredential = "demo@user.io"
+    const demoPassword = "password"
+    return dispatch(sessionActions.login({ credential: demoCredential, password: demoPassword }))
+    .catch(async (res) => {
+      let data;
+      try {
+        // .clone() essentially allows you to read the response body twice
+        data = await res.clone().json();
+      } catch {
+        data = await res.text(); // Will hit this case if the server is down
+      }
+      if (data?.errors) setErrors(data.errors);
+      else if (data) setErrors([data]);
+      else setErrors([res.statusText]);
+    });
+  }
+
+  return (
+    <div className="form-container">
+      <form onSubmit={handleSubmit} className="form">
+        <h1> Log in to continue </h1>
+        <p>Welcome back! Let's get you outside.</p>
+        <ul>
+          {errors.map(error => <li key={error}>{error}</li>)}
+        </ul>
+        <label>
+          <input
+            type="text"
+            value={credential}
+            onChange={(e) => setCredential(e.target.value)}
+            placeholder="Username or Email Address"
+            required
+          />
+        </label>
+        <label>
+          <input
+            type="password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            placeholder="Password"
+            required
+          />
+        </label>
+        <br />
+        <button type="submit">Log In</button>
+        <button type="button" onClick={handleDemoLogin}>Demo User Login</button>
+      </form>
+    </div>
+  );
+}
+
+export default LoginForm;
