+++ frontend/src/components/SessionForms/SessionForms.css
@@ -1,0 +1,103 @@
+#modal {
+    position: fixed;
+    top: 0;
+    right: 0;
+    left: 0;
+    bottom: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-family: Calibre, Helvetica, Arial, sans-serif
+}
+
+#modal-background {
+    position: fixed;
+    top: 0;
+    right: 0;
+    left: 0;
+    bottom: 0;
+    background-color: #313130;
+}
+
+#modal-content {
+    position: absolute;
+    background-color: white;
+    border-radius: 10px;
+}
+
+.modal {
+    border-radius: 10px;
+}
+
+.form-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100%;  
+}
+
+.form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 400px;
+    height: auto;
+    padding: 20px;
+    background-color: rgb(255, 255, 255);
+    border-radius: 10px;
+}
+
+.form label {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    margin: 10px;
+}
+
+.form input {
+    margin: 5px;
+    padding: 5px;
+    border-radius: 5px;
+    border: 1px solid gray;
+    width: 400px;
+    height: 30px;
+}
+
+.form button[type="submit"] {
+    background-color: rgb(0, 197, 133);
+    border-width: 1px;
+    border-style: solid;
+    border-color: rgb(0, 197, 133);
+    color: white;
+    height: 48px;
+    font-size: 18px;
+    border-radius: 8px;
+    display: flex;
+    -webkit-box-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    justify-content: center;
+    padding-left: 24px;
+    padding-right: 24px;
+    flex: 1 1 0%;
+    font-family: Calibre, Helvetica, Arial, sans-serif;
+}
+
+.form h1 {
+    margin-bottom: 20px;
+    text-align:left;
+    color: rgb(40, 64, 12);
+    font-weight: 700;
+    letter-spacing: -0.96px;
+    line-height: 54px;
+}
+
+.form p {
+    color: rgb(60, 60, 60);
+    font-size: 18px;
+    font-weight: 400;
+    letter-spacing: 0px;
+    line-height: 24px;
+}
