+++ frontend/src/components/SessionForms/index.js
@@ -1,0 +1,36 @@
+import React, { useState } from 'react';
+import { Modal } from '../../context/Modal';
+import SignUpForm from '../SessionForms/SignUpForm';
+import LoginForm from '../SessionForms/LoginForm';
+import "./SessionForms.css";
+
+
+export function SignupFormModal() {
+  const [showModal, setShowModal] = useState(false);
+
+  return (
+    <>
+      <button onClick={() => setShowModal(true)} className='signup-nav'>Sign Up</button>
+      {showModal && (
+        <Modal onClose={() => setShowModal(false)}>
+          <SignUpForm />
+        </Modal>
+      )}
+    </>
+  );
+}
+
+export function LoginFormModal() {
+  const [showModal, setShowModal] = useState(false);
+
+  return (
+    <>
+      <button onClick={() => setShowModal(true)} className='login-nav'>Log In</button>
+      {showModal && (
+        <Modal onClose={() => setShowModal(false)}>
+          <LoginForm />
+        </Modal>
+      )}
+    </>
+  );
+}
