--- frontend/src/components/SignupFormModal/index.js
@@ -1,20 +1,0 @@
-import React, { useState } from 'react';
-import { Modal } from '../../context/Modal';
-import SignUpForm from './SignUpForm';
-
-function SignupFormModal() {
-  const [showModal, setShowModal] = useState(false);
-
-  return (
-    <>
-      <button onClick={() => setShowModal(true)}>Sign Up</button>
-      {showModal && (
-        <Modal onClose={() => setShowModal(false)}>
-          <SignUpForm />
-        </Modal>
-      )}
-    </>
-  );
-}
-
-export default SignupFormModal;
